Add auto-calculated dependencies, including duration-based tasks.
This pushes back tasks automatically when their dependent tasks will not
complete in time.  Right now that simply means starting the day after.

The original start date is also recorded so that we could (in theory)
show the delta between the original start and the new start.

Finally, duration-based tasks have their dates calculated (even if they
are not dependant upon or dependencies of other tasks).