Add auto-implied dependencies for tasks that have the same performer.
As with all "automagic" we'll see if this is helpful or hurtful in the
end but I can't imagine there's an easy way to manually do this, or that
the logic would work if we allowed concurrent work here.

For concurrent work I'm instead aiming at having "sub performers" based
on naming convention.  We'll see...