Streamline error messages and permit wider syntax for performer names.
Performers are always in parens following a task name, so while tasks
themselves can't contain parens or certain other characters, we still
want them to be able to be nested or aggregated, so permit '/'
character.

For performers, permit - (for orgs with - in the name) and permit '.'
(to separate numbered example systems from sub-performer numbers).